update: temp account handler reubicated and delete guest virtual machine handler
