+++ servidor_procesamiento/Procesador/main.go
@@ -184,8 +184,6 @@ func manageServer(r *mux.Router) {
 	// Endpoint para crear un usuario nuevo tempral
 	r.HandleFunc(apiPrefix+"temp-user", handlers.CreateTempUserHandler).Methods("POST") 
 
-	//Endpoint para peticiones de inicio de sesiòn
-	// r.HandleFunc("/json/signin", handlers.UserSignInHandler)
 
 	/*
 		-----------------------
